requirements.txt
python-binance
requests
pyTelegramBotApi